Abstraction of Cell
Add more class
Cell is now abstract
