+++ core/src/com/agaroid/Cell.java
@@ -6,41 +6,29 @@ import com.badlogic.gdx.graphics.g2d.GlyphLayout;
 import com.badlogic.gdx.graphics.g2d.SpriteBatch;
 import com.badlogic.gdx.graphics.glutils.ShapeRenderer;
 
-public class Cell {
+public abstract class Cell {
 	public Cell(SpriteBatch batch,
-			ShapeRenderer shapeRenderer, BitmapFont font,
-			String username, int posX, int posY, int mass) {
+			ShapeRenderer shapeRenderer, int posX, int posY, int mass) {
 		super();
 		this.batch = batch;
-		this.font = font;
 		this.shapeRenderer = shapeRenderer;
-		this.username = username;
 		this.posX = posX;
 		this.posY = posY;
 		this.mass = mass;
-		
-		this.glyphLayout = new GlyphLayout();
-        this.glyphLayout.setText(font, username);
-        
         this.color = new Color((float)Math.random(),(float) Math.random(),(float) Math.random(), 1);
 	}
-
 	SpriteBatch batch;
-	BitmapFont font;
-	ShapeRenderer shapeRenderer;
-	
-	GlyphLayout glyphLayout;
-	
-	private String username;
+    ShapeRenderer shapeRenderer;
+
+    GlyphLayout glyphLayout;
+
+    private String username;
 	int posX = 200;
 	int posY = 200;
 	int mass = 200;
 	
 	Color color;
-	
-	public void batchDraw() {
-		font.draw(batch, username, posX - glyphLayout.width /2, posY + glyphLayout.height /2);
-	}
+
 	
 	public void rendererDraw() {
 		shapeRenderer.setColor(color);
@@ -51,6 +39,8 @@ public class Cell {
 		posX += offsetX;
 		posY += offsetY;
 	}
+
+
 	
 	
 }
